+++ Python/tdw/add_ons/replicant.py
@@ -358,7 +358,7 @@ class Replicant(AddOn):
                                            plan=plan)
 
     def grasp(self, target: int, arm: Arm, angle: Optional[float] = 90, axis: Optional[str] = "pitch",
-              relative_to_hand: bool = True) -> None:
+              relative_to_hand: bool = True, offset: float = None) -> None:
         """
         Grasp a target object.
 
@@ -371,6 +371,7 @@ class Replicant(AddOn):
         :param angle: Continuously (per `communicate()` call, including after this action ends), rotate the the grasped object by this many degrees relative to the hand. If None, the grasped object will maintain its initial rotation.
         :param axis: Continuously (per `communicate()` call, including after this action ends) rotate the grasped object around this axis relative to the hand. Options: `"pitch"`, `"yaw"`, `"roll"`. If None, the grasped object will maintain its initial rotation.
         :param relative_to_hand: If True, the object rotates relative to the hand holding it. If False, the object rotates relative to the Replicant. Ignored if `angle` or `axis` is None.
+        :param offset: Offset the object's position from the Replicant's hand by this distance. If None, a default offset is used: `distance(left_bound_of_object, right_bound_of_object) / 3`.
         """
 
         self.action = Grasp(target=target,
@@ -378,7 +379,8 @@ class Replicant(AddOn):
                             dynamic=self.dynamic,
                             angle=angle,
                             axis=axis,
-                            relative_to_hand=relative_to_hand)
+                            relative_to_hand=relative_to_hand,
+                            offset=offset)
 
     def drop(self, arm: Arm, max_num_frames: int = 100, offset: Union[float, np.ndarray, Dict[str, float]] = 0.1) -> None:
         """