+++ Python/tdw/replicant/actions/grasp.py
@@ -17,7 +17,8 @@ class Grasp(Action):
     When an object is grasped, it is made kinematic. Any objects contained by the object are parented to it and also made kinematic. For more information regarding containment in TDW, [read this](../../../lessons/semantic_states/containment.md).
     """
 
-    def __init__(self, target: int, arm: Arm, dynamic: ReplicantDynamic, angle: Optional[float], axis: Optional[str], relative_to_hand: bool):
+    def __init__(self, target: int, arm: Arm, dynamic: ReplicantDynamic, angle: Optional[float], axis: Optional[str],
+                 relative_to_hand: bool, offset: Optional[float]):
         """
         :param target: The target object ID.
         :param arm: The [`Arm`](../arm.md) value for the hand that will grasp the target object.
@@ -25,6 +26,7 @@ class Grasp(Action):
         :param angle: Continuously (per `communicate()` call, including after this action ends), rotate the the grasped object by this many degrees relative to the hand. If None, the grasped object will maintain its initial rotation.
         :param axis: Continuously (per `communicate()` call, including after this action ends) rotate the grasped object around this axis relative to the hand. Options: `"pitch"`, `"yaw"`, `"roll"`. If None, the grasped object will maintain its initial rotation.
         :param relative_to_hand: If True, the object rotates relative to the hand holding it. If False, the object rotates relative to the Replicant. Ignored if `angle` or `axis` is None.
+        :param offset: Offset the object's position from the Replicant's hand by this distance. If None, a default offset is used: `distance(left_bound_of_object, right_bound_of_object) / 3`.
         """
 
         super().__init__()
@@ -48,6 +50,10 @@ class Grasp(Action):
         If True, the object rotates relative to the hand holding it. If False, the object rotates relative to the Replicant. Ignored if `angle` or `axis` is None.
         """
         self.relative_to_hand: bool = relative_to_hand
+        """:field
+        Offset the object's position from the Replicant's hand by this distance. If None, a default offset is used: `distance(left_bound_of_object, right_bound_of_object) / 3`.
+        """
+        self.offset: Optional[float] = offset
         # We're already holding an object.
         if self.arm in dynamic.held_objects:
             self.status = ActionStatus.already_holding
@@ -91,6 +97,12 @@ class Grasp(Action):
                          {"$type": "enable_nav_mesh_obstacle",
                           "id": self.target,
                           "enable": False}])
+        # Set the object's offset.
+        if self.offset is not None:
+            commands.append({"$type": "replicant_set_grasped_object_offset",
+                             "id": static.replicant_id,
+                             "arm": self.arm.name,
+                             "offset": self.offset})
         # Set the object's rotation.
         if self.angle is not None and self.axis is not None:
             commands.append({"$type": "replicant_set_grasped_object_rotation",